Minor package refactor and code re-sturcture
