+++ app/src/main/java/com/jayasuryat/minesweeperjc/presentation/Screen.kt
@@ -1,4 +1,4 @@
-package com.jayasuryat.minesweeperjc.features
+package com.jayasuryat.minesweeperjc.presentation
 
 sealed interface Screen {
