+++ app/src/main/java/com/jayasuryat/minesweeperjc/theme/Color.kt
@@ -1,4 +1,4 @@
-package com.jayasuryat.minesweeperjc.ui.theme
+package com.jayasuryat.minesweeperjc.theme
 
 import androidx.compose.ui.graphics.Color
