+++ app/src/main/java/com/jayasuryat/minesweeperjc/theme/Shape.kt
@@ -1,4 +1,4 @@
-package com.jayasuryat.minesweeperjc.ui.theme
+package com.jayasuryat.minesweeperjc.theme
 
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.Shapes
