+++ ui-game/src/main/java/com/jayasuryat/uigame/GameScreen.kt
@@ -17,8 +17,8 @@ import com.jayasuryat.minesweeperengine.state.asStatefulGrid
 import com.jayasuryat.minesweeperui.composable.grid.GridLayoutInformation
 import com.jayasuryat.uigame.composable.MinefieldScreen
 import com.jayasuryat.uigame.composable.feedback.GameFeedback
-import com.jayasuryat.uigame.feedback.MusicManager
 import com.jayasuryat.uigame.composable.topbar.GameTopBar
+import com.jayasuryat.uigame.feedback.MusicManager
 import com.jayasuryat.uigame.logic.ActionListener
 import com.jayasuryat.uigame.logic.GameConfiguration
 import com.jayasuryat.util.LogCompositions
